allow custom boolean values
While the MIoT specs define the 'bool' type with values 'true' and
'false', some devices apparently use an 'int' with '0' and '1'
instead...

Don't force the 'bool' type for switches and binary sensors, instead
allow the user to set the values representing the boolean true and false
state using the new 'miot_true' and 'miot_false' config variables.

They're optional and default to 'true' and 'false' to cover the sane
MIoT 'bool' type.

Fixes #3.